+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Include string reference id for javascript consumption.
 - Support device renaming.
+- Support logic slot values.
 
 ### 1.0.0